change errors inital value
before it would try to build a shape matching initalValues
with an empty string as the default value. To do this
it would iterate  recursively through the shape and assign the empty
string. Problems is some sub shapes may represent a value primitive
via a shape. For example, a value representing money would be a shape
of {currency: string, amount: number}. In this case, the intial value of
the currency would be {currency: '', amount: ''}, instead of the of
simply null or ''.